+++ .formatter.exs
@@ -2,11 +2,16 @@ spark_locals_without_parens = [
   action: 1,
   default_language: 1,
   fields: 1,
+  index: 1,
   index_name: 1,
+  label_field: 1,
   language_attribute: 1,
   prefix?: 1,
   rank?: 1,
-  search_text_attribute: 1
+  search_text_attribute: 1,
+  source_type: 1,
+  state_attribute: 1,
+  visible_states: 1
 ]
 
 [